backend/services/auth-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
django-cors-headers==4.3.1
django-redis==5.4.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
gunicorn==21.2.0
celery==5.3.4